week7/main-processor-python/requirements.txt
pyserial
PySimpleGUI